src/main/java/com/pocketful/util/EmailValidator.java
package com.pocketful.util;

import java.util.Objects;
import java.util.regex.Matcher;
import java.util.regex.Pattern;

public abstract class EmailValidator {
    public static Boolean validate(String email) {
        if (Objects.isNull(email)) return false;

        String EMAIL_REGEX =  "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$";
        Pattern EMAIL_PATTERN = Pattern.compile(EMAIL_REGEX);
        Matcher matcher = EMAIL_PATTERN.matcher(email);
        return matcher.matches();
    }
}
